% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.fddml}
\alias{print.fddml}
\title{Print an fddml object}
\usage{
\method{print}{fddml}(x, digits = 4, ...)
}
\arguments{
\item{x}{An \code{fddml} object.}

\item{digits}{Number of significant digits (default 4).}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an fddml object
}
