% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.fddml}
\alias{print.summary.fddml}
\title{Print a summary.fddml object}
\usage{
\method{print}{summary.fddml}(x, ...)
}
\arguments{
\item{x}{A \code{summary.fddml} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a summary.fddml object
}
