% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.fddml}
\alias{summary.fddml}
\title{Summary of an fddml object}
\usage{
\method{summary}{fddml}(object, ...)
}
\arguments{
\item{object}{An \code{fddml} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
An object of class \code{summary.fddml}, printed via \code{\link[=print.summary.fddml]{print.summary.fddml()}}.
}
\description{
Summary of an fddml object
}
